Add NumberFormatter and String formatter
TODO:
-  Test/fix StringFormatter for float
- Add Date type
