+++ base/src/iosMain/kotlin/text/StringUtils.kt
@@ -1,0 +1,29 @@
+/*
+ Copyright 2020 Splendo Consulting B.V. The Netherlands
+
+    Licensed under the Apache License, Version 2.0 (the "License");
+    you may not use this file except in compliance with the License.
+    You may obtain a copy of the License at
+
+      http://www.apache.org/licenses/LICENSE-2.0
+
+    Unless required by applicable law or agreed to in writing, software
+    distributed under the License is distributed on an "AS IS" BASIS,
+    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+    See the License for the specific language governing permissions and
+    limitations under the License.
+
+ */
+
+package com.splendo.kaluga.base.text
+
+import com.splendo.kaluga.base.utils.Locale
+import platform.Foundation.NSString
+import platform.Foundation.lowercaseStringWithLocale
+import platform.Foundation.uppercaseStringWithLocale
+
+actual val lineSeparator = "\n"
+
+actual fun String.lowerCased(locale: Locale): String = (this as NSString).lowercaseStringWithLocale(locale)
+
+actual fun String.upperCased(locale: Locale): String = (this as NSString).uppercaseStringWithLocale(locale)
