--- resources/src/androidLibMain/kotlin/strings/StringUtils.kt
@@ -1,32 +1,0 @@
-/*
- Copyright 2020 Splendo Consulting B.V. The Netherlands
- 
-    Licensed under the Apache License, Version 2.0 (the "License");
-    you may not use this file except in compliance with the License.
-    You may obtain a copy of the License at
- 
-      http://www.apache.org/licenses/LICENSE-2.0
- 
-    Unless required by applicable law or agreed to in writing, software
-    distributed under the License is distributed on an "AS IS" BASIS,
-    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-    See the License for the specific language governing permissions and
-    limitations under the License.
-  
- */
-
-package com.splendo.kaluga.resources.strings
-
-import com.splendo.kaluga.base.ApplicationHolder.Companion.applicationContext
-import com.splendo.kaluga.resources.Locale
-
-actual val lineSeparator = System.lineSeparator()
-
-actual fun String.localized(): String {
-    val id = applicationContext.resources.getIdentifier(this, "string", applicationContext.packageName)
-    return applicationContext.getString(id)
-}
-
-actual fun String.lowerCased(locale: Locale): String = this.toLowerCase(locale)
-
-actual fun String.upperCased(locale: Locale): String = this.toUpperCase(locale)
