+++ settings.gradle.kts
@@ -46,3 +46,4 @@ include(":permissions")
 include(":location")
 include(":androidtesthelper")
 include(":keyboard")
+include(":resources")
